moved startup & shutdown automation to separate Automation library to clean up code & create place for future automation
