+++ src/E1_coldflow/config.h
@@ -3,6 +3,7 @@
 #include <Cryo_Thermocouple.h>
 #include "common_fw.h"
 #include <ADS1219.h>
+#include <Automation.h>
 
 
 #define FLIGHT_BRAIN_ADDR 0x00
@@ -50,37 +51,6 @@ const float batteryMonitorShuntR = 0.002; // ohms
 const float batteryMonitorMaxExpectedCurrent = 10; // amps
 
 
-//---------Startup & Shutdown-------------
-
-bool startup = false;
-int startupPhase = 0;
-
-/* Delays during startup sequence:
-  1 - Between open pressure and open LOX Main
-  2 - Between open LOX Main and open Prop Main
-  3 - Between open Prop Main and close Main Valve Arm
-*/
-int startupDelays[3] = {1000, 0 ,1000};
-uint32_t startupTimer;
-
-bool shutdown = false;
-int shutdownPhase = 0;
-
-/* Delays during shutdown sequence:
-  1 - Between arm 2-way, close high pressure, close Prop and close LOX
-  2 - Between close LOX and closing Arming valve & opening Gems vent
-*/
-int shutdownDelays[2] = {0, 750};
-uint32_t shutdownTimer;
-
-int beginFlow();
-void flowConfirmation(float *data);
-int endFlow();
-bool checkStartupProgress(int startupPhase, int startupTimer);
-bool advanceStartup(int startupPhase);
-
-
-
 
 namespace config {
   void setup() {
@@ -99,7 +69,6 @@ namespace config {
 
     debug("Initializing sensors", DEBUG);
     sensors = new sensorInfo[numSensors];
-    // sensorInfo s = ;
     sensors[0] = {"Temperature",   FLIGHT_BRAIN_ADDR, 0, 3}; //&(testTempRead)}, //&(Thermocouple::readTemperatureData)},
     sensors[1] = {"All Pressure",  FLIGHT_BRAIN_ADDR, 1, 1};
     sensors[2] = {"Battery Stats", FLIGHT_BRAIN_ADDR, 2, 3};
@@ -116,7 +85,7 @@ namespace config {
     valves[6] = {"High Pressure Solenoid", 26, &(Solenoids::activateHighPressureSolenoid), &(Solenoids::deactivateHighPressureSolenoid), &(Solenoids::getAllStates)};
     valves[7] = {"Arm Rocket", 27, &(Solenoids::armAll), &(Solenoids::disarmAll), &(Solenoids::getAllStates)};
     valves[8] = {"Launch Rocket", 28, &(Solenoids::LAUNCH), &(Solenoids::endBurn), &(Solenoids::getAllStates)};
-    valves[9] = {"Perform Flow", 29, &(beginFlow), &(endFlow), &(flowConfirmation)};
+    valves[9] = {"Perform Flow", 29, &(Automation::beginFlow), &(Automation::endFlow), &(Automation::flowConfirmation)};
 
     pinMode(LOX_2_PIN, OUTPUT);
     pinMode(LOX_5_PIN, OUTPUT);
@@ -131,140 +100,4 @@ namespace config {
     pinMode(LOX_ADAPTER_HEATER_PIN, OUTPUT);
 
   }
-}
-
-
-// Pretending is a valve action, do -1 to indicate that 
-int beginFlow() {
-  /* Check if rocket is in required state for a flow:
-    Pressure - Closed
-    LOX GEMS & Prop GEMS - Open
-    Arming Valve - Closed
-    LOX Main Valve & Prop Main Valve - Closed
-  */
-  //&& Solenoids::getLoxGems() && Solenoids::getPropGems()
-  startup = !Solenoids::getHPS() &&
-      !Solenoids::getLox2() && !Solenoids::getLox5() && !Solenoids::getProp5();
-  if (startup) {
-     Serial.println("Eureka-1 is in Startup");
-  }
-  return -1;
-}
-
-void flowConfirmation(float *data) {
-  data[0] = startup ? 1 : 0;
-  data[1] = shutdown ? 1 : 0;
-  data[2] = -1;
-}
-
-// Pretending is a valve action, do -1 to indicate that 
-int endFlow() {
-  shutdown = true;
-  Serial.println("Eureka-1 is in Shutdown");
-  return 1;
-}
-
-bool checkStartupProgress(int startupPhase, int startupTimer) {
-  if (startupPhase > 0) {
-    if (startupTimer > startupDelays[startupPhase - 1]) {
-      return true;
-    } else {
-      return false;
-    }
-  } else {
-    return true;
-  }
-}
-
-/* 
-  Progress startup given the current startup phase
-*/
-bool advanceStartup() {
-  if (startupPhase == 0) {
-
-        //close both GEMS, open Arming valve, open Pressurant
-        Solenoids::closeLOXGems();
-        Solenoids::closePropaneGems();
-        Solenoids::armLOX();
-        Solenoids::activateHighPressureSolenoid();
-
-        Solenoids::getAllStates(farrbconvert.sensorReadings);
-        startupPhase++;
-
-  } else if (startupPhase == 1) {
-    
-      // after a delay open LOX main valve
-      Solenoids::openLOX();
-      Solenoids::openPropane();
-      Solenoids::getAllStates(farrbconvert.sensorReadings);
-      startupPhase++;
-
-  } else if (startupPhase == 2) {
-    
-      // after a delay open Prop main valve
-      // Solenoids::openPropane();
-      Solenoids::getAllStates(farrbconvert.sensorReadings);
-      startupPhase++;
-
-  } else if (startupPhase == 3) {
-    
-      // after a delay close Arming valve
-      Solenoids::disarmLOX();
-      Solenoids::getAllStates(farrbconvert.sensorReadings);
-
-
-      startupPhase = 0;
-      startup = false;
-
-  }
-  else {
-    return false;
-  }
-  return true;
-}
-
-
-void shutdownConfirmation(float *data) {
-  data[0] = shutdown ? 1 : 0;
-  data[1] = -1;
-}
-
-bool checkShutdownProgress(int shutdownPhase, int shutdownTimer) {
-  if (shutdownPhase > 0) {
-    if (shutdownTimer > shutdownDelays[shutdownPhase - 1]) {
-      return true;
-    } else {
-      return false;
-    }
-  } else {
-    return true;
-  }
-}
-
-bool advanceShutdown() {
-  if (shutdownPhase == 0) {
-    Solenoids::deactivateHighPressureSolenoid();
-    Solenoids::armLOX();
-    Solenoids::closePropane();
-    Solenoids::getAllStates(farrbconvert.sensorReadings);
-    shutdownPhase++;
-
-  } else if (shutdownPhase == 1) {
-    Solenoids::closeLOX();
-    Solenoids::getAllStates(farrbconvert.sensorReadings);
-    shutdownPhase++;
-
-  } else if (shutdownPhase == 2) {
-    Solenoids::disarmLOX();
-    Solenoids::ventLOXGems();
-    Solenoids::ventPropaneGems();
-    Solenoids::getAllStates(farrbconvert.sensorReadings);
-
-    shutdownPhase = 0;
-    shutdown = false;
-
-  } else {
-    return false;
-  }
-  return true;
 }
